Move telemetry custom screens from telemetry tab to own tab
