+++ companion/src/modeledit/telemetry_customscreen.ui
@@ -31,7 +31,7 @@
         </size>
        </property>
        <property name="text">
-        <string>Custom Screen Type</string>
+        <string>Type</string>
        </property>
       </widget>
      </item>
@@ -123,6 +123,19 @@
     <widget class="QGroupBox" name="screenNums">
      <layout class="QGridLayout" name="screenNumsLayout"/>
     </widget>
+   </item>
+   <item>
+    <spacer name="verticalSpacer">
+     <property name="orientation">
+      <enum>Qt::Vertical</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>20</width>
+       <height>40</height>
+      </size>
+     </property>
+    </spacer>
    </item>
   </layout>
  </widget>
